+++ src/main/java/com/smth/repositories/PersonRepository.java
@@ -1,6 +1,6 @@
 package com.smth.repositories;
 
-import com.smth.domain.Person;
+import com.smth.entities.Person;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PersonRepository extends CrudRepository<Person, Long> {
